Agentic_AI/phidata-Tut/requirements.txt
phidata
groq
numpy
pandas
scikit-learn
yfinance